move plan visitor and rewriter to their own directory
