+++ src/frontend/src/optimizer/heuristic_optimizer.rs
