+++ src/frontend/src/optimizer/mod.rs
@@ -17,45 +17,40 @@ pub use plan_node::PlanRef;
 pub mod property;
 
 mod delta_join_solver;
-mod heuristic;
-mod max_one_row_visitor;
-mod plan_correlated_id_finder;
+mod heuristic_optimizer;
 mod plan_rewriter;
+pub use plan_rewriter::PlanRewriter;
 mod plan_visitor;
 pub use plan_visitor::PlanVisitor;
 mod optimizer_context;
 mod rule;
-mod share_parent_counter;
-mod share_source_rewriter;
 
 use fixedbitset::FixedBitSet;
 use itertools::Itertools as _;
 pub use optimizer_context::*;
+use plan_rewriter::ShareSourceRewriter;
 use property::Order;
 use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::error::{ErrorCode, Result};
 
-use self::heuristic::{ApplyOrder, HeuristicOptimizer};
+use self::heuristic_optimizer::{ApplyOrder, HeuristicOptimizer};
 use self::plan_node::{
     BatchProject, Convention, LogicalProject, StreamDml, StreamMaterialize, StreamRowIdGen,
     StreamSink,
 };
 use self::plan_visitor::{
-    has_batch_exchange, has_batch_seq_scan, has_batch_seq_scan_where, has_logical_apply,
-    has_logical_over_agg,
+    has_batch_exchange, has_batch_seq_scan, has_batch_seq_scan_where, has_batch_source,
+    has_logical_apply, has_logical_over_agg, HasMaxOneRowApply,
 };
 use self::property::RequiredDist;
 use self::rule::*;
 use crate::catalog::column_catalog::ColumnCatalog;
 use crate::catalog::table_catalog::TableType;
 use crate::handler::create_table::DmlFlag;
-use crate::optimizer::max_one_row_visitor::HasMaxOneRowApply;
 use crate::optimizer::plan_node::{
     BatchExchange, ColumnPruningContext, PlanNodeType, PredicatePushdownContext,
 };
-use crate::optimizer::plan_visitor::has_batch_source;
 use crate::optimizer::property::Distribution;
-use crate::optimizer::share_source_rewriter::ShareSourceRewriter;
 use crate::utils::Condition;
 use crate::WithOptions;
