+++ src/frontend/src/optimizer/plan_node/col_pruning.rs
@@ -18,7 +18,7 @@ use paste::paste;
 
 use super::*;
 pub use crate::expr::CollectInputRef;
-use crate::optimizer::share_parent_counter::ShareParentCounter;
+use crate::optimizer::plan_visitor::ShareParentCounter;
 use crate::optimizer::PlanVisitor;
 use crate::{for_batch_plan_nodes, for_stream_plan_nodes};
