+++ src/frontend/src/optimizer/plan_visitor/max_one_row_visitor.rs
