+++ src/frontend/src/optimizer/plan_visitor/share_parent_counter.rs
