+++ src/frontend/src/optimizer/rule/apply_join_transpose_rule.rs
@@ -23,10 +23,10 @@ use crate::expr::{
     CollectInputRef, CorrelatedId, CorrelatedInputRef, Expr, ExprImpl, ExprRewriter, ExprType,
     ExprVisitor, FunctionCall, InputRef,
 };
-use crate::optimizer::plan_correlated_id_finder::{ExprCorrelatedIdFinder, PlanCorrelatedIdFinder};
 use crate::optimizer::plan_node::{
     LogicalApply, LogicalFilter, LogicalJoin, LogicalProject, PlanTreeNodeBinary,
 };
+use crate::optimizer::plan_visitor::{ExprCorrelatedIdFinder, PlanCorrelatedIdFinder};
 use crate::optimizer::PlanRef;
 use crate::utils::{ColIndexMapping, Condition};
